Merged PR 1640: No more NaN values in the frontend for tests
# Implementation Summary
-------------------------------
The test core has been sending NaN values to the frontend in some situations. The MNIST small is an example of a dataset that can produce NaN values for the confusion matrix tests, it's caused by some categories not having any predictions when computing and then normalizing the results for the confusion matrix. To fix it, I simply call the numpy method `nan_to_num()` on the confusion matrix.

# [Optional] How to review
-------------------------------
1. Run MNIST small classification model
2. Run confusion matrix test

# [Optional] Please review the following files
-------------------------------
All

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Added a comment above my fix to explain why it's needed. Small documentation like that goes a long way : )