Merged PR 1634: Large Datasets Slowdown
# Implementation Summary
-------------------------------
This PR seeks to fix the slowdown from #2517, which ultimately is caused by the LW core having to instantiate every layer individually, and running it with input obtained from the output of the previous connected layer in order to get the previews. While this is mostly because of the way Keras works, I suggested downsampling the input so that the LWCore can run it in a more reasonable amount of time, although this creates a domino effect for all layers, where the input for every subsequent layer after the original layer where downsampling was necessary also becomes downsampled.

Because of that, this PR does a small fix instead, where downsampling is in effect only for the previews. This functionality was already there, but there's a lot of evidence to suggest that downsampling the previews doesn't _really_ effect the overall smoothness of the model recommendation -> model workspace transition. However, because downsampling in the LW core is so dangerous, this small change is the next best thing.

# [Optional] How to review
-------------------------------
1. Load the dataset in the bug ticket linked above and make a model with it
2. See how long the transition takes from model recommendation -> workspace/previews loading

# [Optional] Please review the following files
-------------------------------
All

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Wrote a method that could be useful down the line, even though it's not used right now. Included comments to ensure that the pitfalls of that method are addressed when we _do_ decide to take that route to fix this bug.

Related work items: #2517