Merged PR 1660: Reverting confusion matrix to using integers
# Implementation Summary
-------------------------------
After planning more for the new test view, I realized that some important information gets lost when we normalize the confusion matrix. Specifically, what we do now is creating a "recall" confusion matrix, making it impossible to derive the precision.
Later on we will have 3 options for the user; "Standard", "Precision" and "Recall" confusion matrices, but for now we will just go with the standard one.

# [Optional] How to review
-------------------------------
Run the confusion matrix test and see the absolute values rather than fractions.