Merged PR 1651: Training Slowing Each Epoch
# Implementation Summary
-------------------------------
While we've determined that training doesn't slow each epoch from my experiments mapping training time as a function of epochs and mapping the size of latest-training-results.pkl throughout the course of training (pending), the changes made in this branch are significant enough and useful enough to justify publishing a pull request.

Information relating to the size of the latest training results pickle file is added to the application logger. Tracking the size of that file is useful because if it's too large it's potentially problematic.

# [Optional] How to review
-------------------------------
1. Train a model in dev
2. Look for the message 'Size of latest training results in bytes' and 'Size of state pickle file in bytes' in kernel.log

# [Optional] Please review the following files
-------------------------------
All

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Added useful info to kernel logs

Related work items: #2493