Merged PR 1667: Encapsulates how model.jsons are read/written/created within the frontend
# Implementation Summary
-------------------------------
The model.json is a snapshot of the frontend state, which makes it hard for the Kernel to use it. This PR encapsulates how it's created/read from Rygg so that the underlying storage can be changed w/o affecting the frontend in a dramatic way.

This PR does not refactor model.json itself, but prepares by decoupling it from the frontend.

Specifically, I've added helper functions in _/frontend/src/renderer/core/helpers/model-helper.js_ and make sure they're used throughout the codebase.

# [Optional] How to review
-------------------------------

Verify the following:

* Models created in previous versions still work
* Models still work after re-starting the tool
* Saving model.json work as usual
* Training/testing/serving

# [Optional] Please review the following files
-------------------------------
N/A

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------

- Removed lots of dead code.
- Broke out duplicated code into a single function.
- Make the default model consistent across the frontend
- Make sure all models inserted into Rygg are equal (i.e., remove charts in every case)

# Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #2557, #2612