Merged PR 1698: Gets model summary before training starts
# Implementation Summary
-------------------------------
Adds a function that shows the model summary before training starts. We have to run the model on a single sample in order to get the summary because we use subclassed Keras layers to make models in the tool.

# [Optional] How to review
-------------------------------
Run any model, look for a model summary in your terminal tab where the kernel is running.

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
* Made a function that logs the model summary, similar to what we have for epoch summary
* Removed duplicate function definition

Related work items: #2629