Merged PR 1701: Errors in custom components no longer throw an error from the previews
# Implementation Summary
-------------------------------
A quick change was made to `createDataObject.py` to ensure that when a mistake is made in a custom component, such as in this [video](https://perceptilabsgroup.slack.com/files/UP2HCH44D/F02P1P8B5KN/media1.mp4) the tool won't show the error from the `max()` function.

# [Optional] How to review
-------------------------------
Try reproducing the error in the video

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Improving error handling so flow of execution isn't broken. The codebase now throws the correct error, which is reflected in the tool

Related work items: #2574