Merged PR 1793: Gradio now shows predictions as percentages rather than raw values when post processing is disabled
# Implementation Summary
-------------------------------
To allow users to see their predictions as percentages/probabilities rather than a raw value of vectors as it was before, changes were made to the gradio serving module. While it makes sense since that's what the inference model outputs, the raw vector of probabilities on its own is not useful to the user. This story seeks to add value to that information, by displaying this vector of probabilities as percentages instead. So rather than:

![image.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1784/attachments/image.png) https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/wit/attachments/805ce50a-2dd4-4174-90e6-b357ff366e3b?fileName=Screen%20Shot%202022-01-20%20at%2011.17.18%20AM.png

We now have:
![Screen Shot 2022-01-26 at 12.29.48 PM.png](https://dev.azure.com/PerceptiLabs/54ade330-1468-4064-8c98-7fe751f97330/_apis/git/repositories/633afcd6-50cc-4639-adbe-75977ad6b69a/pullRequests/1784/attachments/Screen%20Shot%202022-01-26%20at%2012.29.48%20PM.png)

# [Optional] How to review
-------------------------------
Try serving a model with categorical outputs and image inputs with post-processing enabled and disabled. With it enabled, it should output the category itself, which would be a string. With it disabled, it should output the vector, displayed as percentages by gradio.

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Added helper methods to get the dtype and categories as strings for more readable code.

Related work items: #2724