Merged PR 1819: Allows for interactive debugger for kernel when using docker version of tool
# Implementation Summary
-------------------------------
A small change to allow for an interact-able debugger. For info on how to debug when developing for the docker version, see this [notion page](https://www.notion.so/How-to-use-the-python-debugger-when-developing-for-the-cd-version-locally-dab408a0ff764429b9f12a33a1e6385a).

# [Optional] How to review
-------------------------------
Set a breakpoint in the kernel after following the steps in the notion page.

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Added a convenience for the other devs on the team hopefully