Merged PR 1827: Examples now show in Gradio
# Implementation Summary
-------------------------------
Examples are taken from the dataset that the model we're deploying comes from. They're then saved to a temporary directory since examples only work with Gradio when they're saved in the same relative path as the script that starts the Gradio interface.

# [Optional] How to review
-------------------------------
Try deploying a model with image inputs (so classification or segmentation should work) and some examples from the test partition of the dataset should show up.

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Kept code clean by using methods to create the examples and delete the temp directory created to store the examples.

Made improvements durable by not including examples when preprocessing/postprocesing isn't included in deployment, since the sample images don't work as valid inputs to the served model without them.

Related work items: #2494