Merged PR 1859: Testing no longer hangs in Docker build
# Implementation Summary
-------------------------------
Tests were hanging in the docker version because files that get stored in temporary directories aren't preserved between containers. For this bug, the docker container `dev-training-1` would save testing results to a temporary directory, and the container `dev-render-1` would try to read from that temporary directory to decide what gets shown in the frontend. However a file that gets created in the former container is not also created in the latter. To work around this, we now leverage the Redis Cache for storing information relating to the status of tests started by the user as a fallback option if the file does not contain the correct information/does not exist.

# [Optional] How to review
-------------------------------
* Train a model and test it locally
* Train a model and test it on Docker (go to `perceptilabs/docker/dev`, start docker desktop, and run `docker-compose down` and `docker-compose up` to get started with the local docker version)

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
Switched `os.path` to `Pathlib` since the `os` submodule treats paths as strings vs. `Pathlib` which treats them as objects.

Related work items: #2728