Merged PR 1863: Groundwork for shap values
# Implementation Summary
-------------------------------
This PR lays out the groundwork for the shap values test. The intention is to get some early feedback on it before proceeding with the rest of the feature.

Note: no results are shown in the frontend. They're written to your home directory If that's not available (docker version) you can set the full path via the environment variable PL_SHAP_PATH.

As for the implementation, it contains work on shapitself , but also some changes to the testcore and testing interfaces to allow for easier unit testing and to support tests that don't rely on us making the model predictions.

# [Optional] How to review
-------------------------------
1. Run the shap values test on a single-input/output classification model.
2. Check your home directory for a file called shap_plot.png (the exact path is also written to the console)

# [Optional] Please review the following files
-------------------------------
N/A

# What did you do to make the codebase a better place? How did you make the improvements durable
-------------------------------
*Here you should describe what improvements you made to codebase while working on the ticket.*
*Note that the improvements should be proportional in size to the main objective of the ticket*

# Please also:
- [ ] Include any new files in included_files.txt
- [ ] Update the story on the board
- [ ] Announce changes that will affect everyone in #dev in slack

Related work items: #2878